various size_t->guint64 fixes
some platforms have off_t as 32-bits breaking large file support ...
make sure we use guint64 everywhere

+++ libvips/include/vips/header.h
@@ -77,7 +77,7 @@ extern "C" {
  */
 #define VIPS_META_RESOLUTION_UNIT "resolution-unit"
 
-int vips_format_sizeof( VipsBandFormat format );
+guint64 vips_format_sizeof( VipsBandFormat format );
 
 int vips_image_get_width( const VipsImage *image );
 int vips_image_get_height( const VipsImage *image );
